apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "fastapi[standard]==0.116.*",
    "openai==1.93.*",
    "pydantic-settings==2.10.*",
    "python-socketio==5.13.*",
]
